Refactored entire script with functions
Refactored to use functions and it works more efficiently to get carrier/porting info for all numbers by breaking up the phone numbers into the list before even starting the API call. Plus, updated code is able to get remaining phone numbers that it previously was not able to. 

Also, the merge of DataFrames is now working. Had to make the 'Business Phone' columns the same type (int).